+++ near-sdk/src/environment/env.rs
@@ -7,10 +7,10 @@
 //! In case of cross-contract calls prefer using higher-level API available
 //! through [`crate::Promise`], and [`crate::PromiseOrValue<T>`].
 
+use std::convert::TryFrom;
 use std::convert::TryInto;
 use std::mem::{size_of, size_of_val};
 use std::panic as std_panic;
-use std::{convert::TryFrom, mem::MaybeUninit};
 
 #[cfg(all(not(target_arch = "wasm32"), feature = "unit-testing"))]
 use crate::mock::MockedBlockchain;
@@ -21,6 +21,9 @@ use crate::types::{
     AccountId, BlockHeight, Gas, NearToken, PromiseIndex, PromiseResult, PublicKey, StorageUsage,
 };
 use crate::{CryptoHash, GasWeight, PromiseError};
+
+#[cfg(feature = "deterministic-account-ids")]
+use crate::{AccountContract, ActionIndex};
 use near_sys as sys;
 
 const REGISTER_EXPECTED_ERR: &str =
@@ -60,25 +63,10 @@ macro_rules! method_into_register {
     }};
 }
 
-//* Note: need specific length functions because const generics don't work with mem::transmute
-//* https://github.com/rust-lang/rust/issues/61956
-
-pub(crate) unsafe fn read_register_fixed_20(register_id: u64) -> [u8; 20] {
-    let mut hash = [MaybeUninit::<u8>::uninit(); 20];
-    sys::read_register(register_id, hash.as_mut_ptr() as _);
-    std::mem::transmute(hash)
-}
-
-pub(crate) unsafe fn read_register_fixed_32(register_id: u64) -> [u8; 32] {
-    let mut hash = [MaybeUninit::<u8>::uninit(); 32];
-    sys::read_register(register_id, hash.as_mut_ptr() as _);
-    std::mem::transmute(hash)
-}
-
-pub(crate) unsafe fn read_register_fixed_64(register_id: u64) -> [u8; 64] {
-    let mut hash = [MaybeUninit::<u8>::uninit(); 64];
-    sys::read_register(register_id, hash.as_mut_ptr() as _);
-    std::mem::transmute(hash)
+pub(crate) unsafe fn read_register_fixed<const N: usize>(register_id: u64) -> [u8; N] {
+    let mut buf = [0; N];
+    sys::read_register(register_id, buf.as_mut_ptr() as _);
+    buf
 }
 
 /// Replaces the current low-level blockchain interface accessible through `env::*` with another
@@ -169,6 +157,41 @@ pub fn register_len(register_id: u64) -> Option<u64> {
 /// ```
 pub fn current_account_id() -> AccountId {
     assert_valid_account_id(method_into_register!(current_account_id))
+}
+
+/// The code of the current contract.
+///
+/// # Examples
+/// ```
+/// use near_sdk::env::current_contract_code;
+/// use near_sdk::AccountContract;
+///
+/// assert!(matches!(current_contract_code(), AccountContract::Local(_)));
+/// ```
+#[cfg(feature = "deterministic-account-ids")]
+pub fn current_contract_code() -> AccountContract {
+    let mode = unsafe { sys::current_contract_code(ATOMIC_OP_REGISTER) };
+    match mode {
+        0 => AccountContract::None,
+        1 => AccountContract::Local(unsafe { read_register_fixed(ATOMIC_OP_REGISTER) }),
+        2 => AccountContract::Global(unsafe { read_register_fixed(ATOMIC_OP_REGISTER) }),
+        3 => AccountContract::GlobalByAccount(assert_valid_account_id(method_into_register!(
+            current_account_id
+        ))),
+        _ => panic!("Invalid contract mode"),
+    }
+}
+
+/// The account id that will receive the refund if the contract panics.
+///
+/// # Examples
+/// ```
+///
+/// assert_eq!(refund_to_account_id(), "bob.near".parse::<AccountId>().unwrap());
+/// ```
+#[cfg(feature = "deterministic-account-ids")]
+pub fn refund_to_account_id() -> AccountId {
+    assert_valid_account_id(method_into_register!(refund_to_account_id))
 }
 
 /// The id of the account that either signed the original transaction or issued the initial
@@ -458,7 +481,7 @@ pub fn random_seed_array() -> [u8; 32] {
     //*         because all bytes are filled. This assumes a valid random_seed implementation.
     unsafe {
         sys::random_seed(ATOMIC_OP_REGISTER);
-        read_register_fixed_32(ATOMIC_OP_REGISTER)
+        read_register_fixed(ATOMIC_OP_REGISTER)
     }
 }
 
@@ -533,7 +556,7 @@ pub fn sha256_array(value: impl AsRef<[u8]>) -> CryptoHash {
     //*         because all bytes are filled. This assumes a valid sha256 implementation.
     unsafe {
         sys::sha256(value.len() as _, value.as_ptr() as _, ATOMIC_OP_REGISTER);
-        read_register_fixed_32(ATOMIC_OP_REGISTER)
+        read_register_fixed(ATOMIC_OP_REGISTER)
     }
 }
 
@@ -558,7 +581,7 @@ pub fn keccak256_array(value: impl AsRef<[u8]>) -> CryptoHash {
     //*         because all bytes are filled. This assumes a valid keccak256 implementation.
     unsafe {
         sys::keccak256(value.len() as _, value.as_ptr() as _, ATOMIC_OP_REGISTER);
-        read_register_fixed_32(ATOMIC_OP_REGISTER)
+        read_register_fixed(ATOMIC_OP_REGISTER)
     }
 }
 
@@ -583,7 +606,7 @@ pub fn keccak512_array(value: impl AsRef<[u8]>) -> [u8; 64] {
     //*         because all bytes are filled. This assumes a valid keccak512 implementation.
     unsafe {
         sys::keccak512(value.len() as _, value.as_ptr() as _, ATOMIC_OP_REGISTER);
-        read_register_fixed_64(ATOMIC_OP_REGISTER)
+        read_register_fixed(ATOMIC_OP_REGISTER)
     }
 }
 
@@ -608,7 +631,7 @@ pub fn ripemd160_array(value: impl AsRef<[u8]>) -> [u8; 20] {
     //*         because all bytes are filled. This assumes a valid ripemd160 implementation.
     unsafe {
         sys::ripemd160(value.len() as _, value.as_ptr() as _, ATOMIC_OP_REGISTER);
-        read_register_fixed_20(ATOMIC_OP_REGISTER)
+        read_register_fixed(ATOMIC_OP_REGISTER)
     }
 }
 
@@ -639,7 +662,7 @@ pub fn ecrecover(
         if return_code == 0 {
             None
         } else {
-            Some(read_register_fixed_64(ATOMIC_OP_REGISTER))
+            Some(read_register_fixed(ATOMIC_OP_REGISTER))
         }
     }
 }
@@ -1087,6 +1110,134 @@ pub fn promise_batch_then(promise_index: PromiseIndex, account_id: &AccountId) -
     }
 }
 
+/// Set the account id that will receive the refund if the promise panics.
+/// Uses low-level [`crate::env::promise_set_refund_to`]
+/// # Examples
+/// ```
+/// use near_sdk::env::{promise_set_refund_to, promise_create};
+/// use near_sdk::AccountId;
+/// use std::str::FromStr;
+///
+/// let promise = promise_create(
+///     "account.near".parse().unwrap(),
+///     "method",
+///     [],
+///     NearToken::from_millinear(1),
+///     NearGas::from_tgas(1),
+/// );
+/// promise_set_refund_to(promise, "refund.near".parse().unwrap());
+/// ```
+#[cfg(feature = "deterministic-account-ids")]
+pub fn promise_set_refund_to(promise_index: PromiseIndex, account_id: &AccountId) {
+    let account_id: &str = account_id.as_ref();
+    unsafe {
+        sys::promise_set_refund_to(promise_index.0, account_id.len() as _, account_id.as_ptr() as _)
+    }
+}
+
+/// Appends `DeterministicStateInit` action to the batch of actions for the given promise
+/// pointed by `promise_index`
+/// Uses low-level [`crate::env::promise_batch_action_state_init`]
+/// # Examples
+/// ```
+/// use near_sdk::env::{promise_batch_action_state_init, promise_create};
+/// use near_sdk::AccountId;
+/// use std::str::FromStr;
+///
+/// let promise = promise_create(
+///     "account.near".parse().unwrap(),
+///     "method",
+///     [],
+///     NearToken::from_millinear(1),
+///     NearGas::from_tgas(1),
+/// );
+/// promise_batch_action_state_init(promise, CryptoHash::from_str("code_hash").unwrap(), NearToken::from_millinear(1));
+#[cfg(feature = "deterministic-account-ids")]
+pub fn promise_batch_action_state_init(
+    promise_index: PromiseIndex,
+    code: CryptoHash,
+    amount: NearToken,
+) -> ActionIndex {
+    unsafe {
+        sys::promise_batch_action_state_init(
+            promise_index.0,
+            code.len() as _,
+            code.as_ptr() as _,
+            &amount.as_yoctonear() as *const u128 as _,
+        )
+    }
+}
+
+/// Appends `DeterministicStateInit` action to the batch of actions for the given promise
+/// pointed by `promise_index`
+/// Uses low-level [`crate::env::promise_batch_action_state_init_by_account_id`]
+/// # Examples
+/// ```
+/// use near_sdk::env::{promise_batch_action_state_init_by_account_id, promise_create};
+/// use near_sdk::AccountId;
+/// use std::str::FromStr;
+///
+/// let promise = promise_create(
+///     "account.near".parse().unwrap(),
+///     "method",
+///     [],
+///     NearToken::from_millinear(1),
+///     NearGas::from_tgas(1),
+/// );
+/// promise_batch_action_state_init_by_account_id(promise, "account.near".parse().unwrap(), NearToken::from_millinear(1));
+/// ```
+#[cfg(feature = "deterministic-account-ids")]
+pub fn promise_batch_action_state_init_by_account_id(
+    promise_index: PromiseIndex,
+    account_id: &AccountIdRef,
+    amount: NearToken,
+) -> ActionIndex {
+    unsafe {
+        sys::promise_batch_action_state_init_by_account_id(
+            promise_index.0,
+            account_id.as_bytes().len() as _,
+            account_id.as_bytes().as_ptr() as _,
+            &amount.as_yoctonear() as *const u128 as _,
+        )
+    }
+}
+
+/// Appends a data entry to an existing `DeterministicStateInit` action.
+/// Uses low-level [`crate::env::set_state_init_data_entry`]
+/// # Examples
+/// ```
+/// use near_sdk::env::{set_state_init_data_entry, promise_create};
+/// use near_sdk::AccountId;
+/// use std::str::FromStr;
+///
+/// let promise = promise_create(
+///     "account.near".parse().unwrap(),
+///     "method",
+///     [],
+///     NearToken::from_millinear(1),
+///     NearGas::from_tgas(1),
+/// );
+/// let action_index = promise_batch_action_state_init_by_account_id(promise, "account.near".parse().unwrap(), NearToken::from_millinear(1));
+/// set_state_init_data_entry(promise, action_index, b"key", b"value");
+/// ```
+#[cfg(feature = "deterministic-account-ids")]
+pub fn set_state_init_data_entry(
+    promise_index: PromiseIndex,
+    action_index: ActionIndex,
+    key: &[u8],
+    value: &[u8],
+) {
+    unsafe {
+        sys::set_state_init_data_entry(
+            promise_index.0,
+            action_index,
+            key.len() as _,
+            key.as_ptr() as _,
+            value.len() as _,
+            value.as_ptr() as _,
+        )
+    }
+}
 /// Attach a create account promise action to the NEAR promise index with the provided promise index.
 ///
 /// More info about batching [here](crate::env::promise_batch_create)